+++ app/app.config.ts
@@ -1,5 +1,6 @@
 import alert from "./config/alert";
 import badge from "./config/badge";
+import breadcrumb from "./config/breadcrumb";
 import button from "./config/button";
 import checkboxGroup from "./config/checkbox-group";
 import colors from "./config/colors";
@@ -29,6 +30,7 @@ export default defineAppConfig({
 		badge,
 		tabs,
 		button,
+		breadcrumb,
 		dashboardPanel,
 		navigationMenu,
 		dashboardNavbar,